Heka should waitForTrigger, it seems
Installed the arduino threshold controller to allow TTL triggering on
the Heka ITC18 TRIG IN port even if the photodiode voltage is low.

Implemented waitForTrigger as property to UvLCRStageProtocol as a way to
select whether to force the heka to wait for the photodiode trigger_in
via epoch.shouldWaitForTrigger.

In same class, added addTrackerBarToFirstFrame method (& modified
addFrameTracker) to enable the tracker bar to be played once at the
onset of the epoch to signal the heka, in addition to the
every-other-frame tracker bar that is currently played during the
stimulus presentation.